requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-api-python-client>=2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 